+++ apps/web/partials/entity-page/editable-entity-page.tsx
@@ -912,8 +912,8 @@ function RenderedValue({
           </>
         );
       }
-      case 'INT64':
-      case 'FLOAT64':
+      case 'INTEGER':
+      case 'FLOAT':
       case 'DECIMAL':
         return (
           <NumberField
